+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
+
 ## v3.0.0
+
 ### Added
+
 - custom headers
 - sequential workload files
-- bucket in json results 
+- bucket in json results
 - addressingStyleVirtual option
 - random range option
 - custom query parameters option
@@ -11,9 +14,10 @@
 - tagging-directive option
 
 ### Changed
+
 - duration option to support other flag settings
 
 ### Removed
+
 - replay option and the previous workload option was renamed to mixed-workload
 - consistency control option
-